backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.9.0.0
